fix theme and lang during state init
