+++ ui/src/layouts/file_preview.rs
@@ -59,16 +59,17 @@ pub fn get_file_format(file_name: String) -> FileFormat {
 #[inline_props]
 #[allow(non_snake_case)]
 pub fn FilePreview(cx: Scope, file: File, _drop_handler: WindowDropHandler) -> Element {
+    let state = use_shared_state::<State>(cx)?;
     let file_format = get_file_format(file.name());
     let file_name = file.name();
     let thumbnail = file.thumbnail();
     let has_thumbnail = !file.thumbnail().is_empty();
     let desktop = use_window(cx);
-    let mut css_style = update_theme_colors();
+    let mut css_style = update_theme_colors(state);
     let update_state: &UseRef<Option<()>> = use_ref(cx, || Some(()));
 
     if update_state.read().is_some() {
-        css_style = update_theme_colors();
+        css_style = update_theme_colors(state);
         *update_state.write_silent() = None;
     }
 
@@ -248,9 +249,9 @@ fn resize_window(
     Some(())
 }
 
-fn update_theme_colors() -> String {
-    let state = State::load();
+fn update_theme_colors(state: UseSharedState<State>) -> String {
     let mut css_style = state
+        .read()
         .ui
         .theme
         .as_ref()
